Split up Game for better dev experience
