+++ Sources/ViewControllers/Game/Game+Haptics.swift
@@ -1,0 +1,29 @@
+//
+//  Game+Haptics.swift
+//  Bondi Ball
+//
+//  Created by Sebastiaan Hols on 12/02/2023.
+//  Copyright © 2023 Dream Emulator. All rights reserved.
+//
+
+import UIKit
+
+// - MARK: Haptics
+extension GameVC {
+  func fingerOnBallHaptic() {
+    let generator = UIImpactFeedbackGenerator(style: .light)
+    generator.impactOccurred()
+  }
+
+  func releaseBallHaptic(withVelocity velocity: CGVector) {
+    if velocity.length > 1 {
+      let generator = UIImpactFeedbackGenerator(style: .medium)
+      generator.impactOccurred()
+    }
+  }
+
+  func ballInPocketHaptic() {
+    let generator = UIImpactFeedbackGenerator(style: .heavy)
+    generator.impactOccurred()
+  }
+}
