+++ Sources/ViewControllers/Game/Game.xib
@@ -16,12 +16,12 @@
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="10" id="6Da-dp-kS9" userLabel="V-Stack">
+        <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="20" id="6Da-dp-kS9" userLabel="V-Stack">
             <rect key="frame" x="0.0" y="0.0" width="183" height="200"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <subviews>
                 <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" maximumZoomScale="2" dataMode="none" id="l4R-Ia-4pN" userLabel="Grid">
-                    <rect key="frame" x="0.0" y="0.0" width="183" height="186"/>
+                    <rect key="frame" x="0.0" y="0.0" width="183" height="176"/>
                     <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                     <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                     <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="Eoo-UZ-eRu">
